refactor(agent): simplify architecture with centralized resolution
- Add AgentIndex.resolve() as the primary resolution method
  - Handles both "source:name" and legacy "name" formats
  - Manages _multiple variant resolution
  - Single source of truth for agent path resolution

- Remove duplicate resolveFromAgentIndex() from executeAgent.ts
  - Now uses AgentIndex.resolve() directly
  - Cleaner, shorter getAgentPath() function

- Merge computeAgentOptions.ts into AgentOptionsBuilder.ts
  - Consolidates sync/async option computation in one place
  - Removes unnecessary wrapper layer
  - Delete computeAgentOptions.ts (code moved to AgentOptionsBuilder)

This reduces abstraction layers and establishes clearer patterns:
- AgentIndex.resolve() for path resolution
- AgentOptionsBuilder for all dropdown option computation
