+++ src/agent/index/index.ts
@@ -19,6 +19,9 @@ export {
   buildAgentOptionsFromIndex,
   getDefaultWorkflowAgent,
   getDefaultToolUseAgent,
+  computeAgentOptions,
+  computeAgentOptionsSync,
+  refreshAgentOptions,
   DEFAULT_WORKFLOW_AGENT,
   DEFAULT_TOOL_USE_AGENT,
   type AgentOptionsPayload,
